Moved register_codecs.kt from helper to helper.codecs
