+++ src/main/kotlin/com/ex_dock/ex_dock/database/scope/ScopeJdbcVerticle.kt
@@ -2,7 +2,6 @@ package com.ex_dock.ex_dock.database.scope
 
 import com.ex_dock.ex_dock.database.connection.getConnection
 import com.ex_dock.ex_dock.frontend.cache.setCacheFlag
-import com.ex_dock.ex_dock.helper.registerGenericCodec
 import com.ex_dock.ex_dock.helper.replyListMessage
 import com.ex_dock.ex_dock.helper.replySingleMessage
 import io.vertx.core.Future
@@ -11,10 +10,6 @@ import io.vertx.core.eventbus.DeliveryOptions
 import io.vertx.core.eventbus.EventBus
 import io.vertx.core.json.JsonObject
 import io.vertx.ext.mongo.MongoClient
-import io.vertx.jdbcclient.JDBCPool
-import io.vertx.sqlclient.Pool
-import io.vertx.sqlclient.Row
-import io.vertx.sqlclient.Tuple
 
 class ScopeJdbcVerticle:  VerticleBase() {
   private lateinit var client: MongoClient
