+++ src/test/kotlin/com/ex_dock/ex_dock/backend/v1/router/sales/SalesRouterTest.kt
@@ -4,7 +4,7 @@ import com.ex_dock.ex_dock.MainVerticle
 import com.ex_dock.ex_dock.database.sales.*
 import com.ex_dock.ex_dock.helper.deployWorkerVerticleHelper
 import com.ex_dock.ex_dock.helper.load
-import com.ex_dock.ex_dock.helper.registerGenericCodec
+import com.ex_dock.ex_dock.helper.codecs.registerGenericCodec
 import io.vertx.core.Future
 import io.vertx.core.Vertx
 import io.vertx.core.eventbus.DeliveryOptions
