+++ src/test/kotlin/com/ex_dock/ex_dock/database/auth/AuthenticationVerticleTest.kt
@@ -3,7 +3,7 @@ package com.ex_dock.ex_dock.database.auth
 import com.ex_dock.ex_dock.database.account.FullUser
 import com.ex_dock.ex_dock.database.account.Permission
 import com.ex_dock.ex_dock.helper.deployWorkerVerticleHelper
-import com.ex_dock.ex_dock.helper.registerGenericCodec
+import com.ex_dock.ex_dock.helper.codecs.registerGenericCodec
 import io.vertx.core.Vertx
 import io.vertx.core.eventbus.DeliveryOptions
 import io.vertx.core.eventbus.EventBus
