+++ src/test/kotlin/com/ex_dock/ex_dock/database/sales/SalesJdbcVerticleTest.kt
@@ -1,8 +1,7 @@
 package com.ex_dock.ex_dock.database.sales
 
 import com.ex_dock.ex_dock.helper.deployWorkerVerticleHelper
-import com.ex_dock.ex_dock.helper.registerGenericCodec
-import io.vertx.core.CompositeFuture
+import com.ex_dock.ex_dock.helper.codecs.registerGenericCodec
 import io.vertx.core.Future
 import io.vertx.core.Vertx
 import io.vertx.core.eventbus.DeliveryOptions
