+++ src/test/kotlin/com/ex_dock/ex_dock/database/url/UrlJdbcVerticleTest.kt
@@ -1,7 +1,7 @@
 package com.ex_dock.ex_dock.database.url
 
 import com.ex_dock.ex_dock.helper.deployWorkerVerticleHelper
-import com.ex_dock.ex_dock.helper.registerGenericCodec
+import com.ex_dock.ex_dock.helper.codecs.registerGenericCodec
 import io.vertx.core.Vertx
 import io.vertx.core.eventbus.DeliveryOptions
 import io.vertx.core.eventbus.EventBus
